+++ tests/e2e/cases/generated_c_dce_const/in.ts
@@ -561,6 +561,12 @@ const unused_promise_try_empty_call = Promise.try(() => {
 const unused_promise_try_resolve_call = Promise.try(() => Promise.resolve("dead_promise_try_resolve"));
 const unused_promise_try_reject_call = Promise.try(() => Promise.reject("dead_promise_try_reject"));
 const unused_promise_try_all_empty_call = Promise.try(() => Promise.all([] as Promise<string>[]));
+const unused_promise_try_then_call = Promise.try(() => "dead_promise_try_then_source")
+    .then(() => "dead_promise_try_then_callback");
+const unused_promise_try_catch_call = Promise.try(() => Promise.reject("dead_promise_try_catch_source"))
+    .catch(() => "dead_promise_try_catch_callback");
+const unused_promise_try_finally_call = Promise.try(() => Promise.resolve("dead_promise_try_finally_source"))
+    .finally(() => "dead_promise_try_finally_callback");
 const unused_new_promise_empty_executor_call = new Promise<string>(() => {
 });
 const unused_new_promise_resolve_call = new Promise<string>((resolve) => resolve("dead_new_promise_resolve"));
@@ -1595,6 +1601,12 @@ Promise.try(() => {
 Promise.try(() => Promise.resolve("top_level_dead_promise_try_resolve"));
 Promise.try(() => Promise.reject("top_level_dead_promise_try_reject"));
 Promise.try(() => Promise.any([] as Promise<string>[]));
+Promise.try(() => "top_level_dead_promise_try_then_source")
+    .then(() => "top_level_dead_promise_try_then_callback");
+Promise.try(() => Promise.reject("top_level_dead_promise_try_catch_source"))
+    .catch(() => "top_level_dead_promise_try_catch_callback");
+Promise.try(() => Promise.resolve("top_level_dead_promise_try_finally_source"))
+    .finally(() => "top_level_dead_promise_try_finally_callback");
 new Promise<string>(() => {
 });
 new Promise<string>((resolve) => resolve("top_level_dead_new_promise_resolve"));
@@ -2606,6 +2618,12 @@ function usedLocal(value: number): number {
     Promise.try(() => Promise.resolve("local_dead_promise_try_resolve"));
     Promise.try(() => Promise.reject("local_dead_promise_try_reject"));
     Promise.try(() => Promise.race([] as Promise<string>[]));
+    Promise.try(() => "local_dead_promise_try_then_source")
+        .then(() => "local_dead_promise_try_then_callback");
+    Promise.try(() => Promise.reject("local_dead_promise_try_catch_source"))
+        .catch(() => "local_dead_promise_try_catch_callback");
+    Promise.try(() => Promise.resolve("local_dead_promise_try_finally_source"))
+        .finally(() => "local_dead_promise_try_finally_callback");
     new Promise<string>(() => {
     });
     new Promise<string>((resolve) => resolve("local_dead_new_promise_resolve"));